requirements.txt
asgiref==3.8.1
Django==4.2.1
django-leaflet==0.29.1
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
